+++ tests/js/realm-tests.js
@@ -2055,11 +2055,11 @@ module.exports = {
 
     TestCase.assertThrowsContaining(() => {
       realm.writeCopyTo();
-    }, "At least path has to be provided for 'writeCopyTo'");
+    }, "`writeCopyTo` requires <output configuration> or <path, [encryptionKey]> parameters");
 
     TestCase.assertThrowsContaining(() => {
       realm.writeCopyTo(34);
-    }, "Argument to 'writeCopyTo' must be a String.");
+    }, "`config` parameter must be an object");
 
     // make sure that copies are in the same directory as the original file
     // that is important for running tests on mobile devices,
@@ -2075,7 +2075,7 @@ module.exports = {
 
     TestCase.assertThrowsContaining(() => {
       realm.writeCopyTo(realm.path + ".copy-invalid-key.realm", "hello");
-    }, "Encryption key for 'writeCopyTo' must be a Binary.");
+    }, "Encryption key for 'writeCopyTo' must be a Binary");
 
     const encryptedCopyName = realm.path + ".copy-encrypted.realm";